Added user based page views
Created groups named 'admin' and 'customer' in /admin page
Assigned one user as amin and others as customer

Created decorators.py to store decorator functions.
unauthenticated_user function doesn't allow logged in users to ente localhost:8000/login or /register and redirects them to the dashboard
allowed_users function only allows 'users, passed as the list in decorator argument' to view certain pages, in this case, admin gets to view all the pages, customer is unauthorized to view pages except dashboard
admin_only function is to fix a problem, when a user refreshes their dashboard, they are sent to main dashboard which they are not authorized to view. Although this isn't the perfect solution, it redirects the user tho the user page and the admin to the dashboard

In navbar.html, two items from the nav bar are hidden for anyone except 'staff'